Integrated Matt's energy bar code. Implemented parallax scrolling riverbanks. Modified spawning.

+++ renderer.py
@@ -12,7 +12,7 @@ class SoftwareRenderer(sdl2.ext.SoftwareSpriteRenderSystem):
         # Game Over Rendering:
         if globals.death == True:
             # Render everyhing except player and enemy sprites:
-            valid = [sprite for sprite in components if sprite.depth not in [2,3] and sprite.depth!=5]
+            valid = [sprite for sprite in components if sprite.depth not in [2,3,5]]
             # Delete all player and enemy sprites:
             delete = set(components) - set(valid)
             self.render(sorted(valid, key=self._sortfunc))
@@ -22,7 +22,8 @@ class SoftwareRenderer(sdl2.ext.SoftwareSpriteRenderSystem):
         # Game Play Rendering:
         elif globals.home_lock == False:
             # Render everything except enemies that have reached the bottom of the screen:
-            valid = [sprite for sprite in components if (sprite.depth==3 and sprite.area[3]==650)==False]
+            valid = [sprite for sprite in components if
+                (sprite.depth==3 and (sprite.area[3]==650 or sprite.area[3]>650))==False]
             self.render(sorted(valid, key=self._sortfunc))
             # Delete the score sprite on each process loop:
             hud_sprites = [sprite for sprite in components if sprite.depth==5]
@@ -38,6 +39,11 @@ class SoftwareRenderer(sdl2.ext.SoftwareSpriteRenderSystem):
             delete = set(components) - set(valid)
             for sprite in delete:
                 entity = world.get_entities(sprite)[0]
+                entity.delete()
+            # Delete the energy bar:
+            energy_bar = [sprite for sprite in components if sprite.depth == 5 and sprite.x==625]
+            if energy_bar:
+                entity = world.get_entities(energy_bar[0])[0]
                 entity.delete()
         # Home Screen Rendering:
         else:
